feat: 'wkndPlanning' dir renamed to 'src' and refactored all references
