+++ src/static/css/manage_mappings.css
