+++ src/static/js/index.js
