+++ src/static/js/manage_mappings_globals.js
