+++ src/static/js/manage_mappings_main.js
