+++ src/static/js/manage_mappings_satellite_lines.js
