+++ src/static/js/manage_mappings_task_technology.js
