+++ src/static/js/manage_mappings_technician_data.js
