+++ src/static/js/manage_mappings_technician_groups.js
