+++ src/static/js/manage_mappings_technician_skills.js
