+++ src/static/js/manage_mappings_technologies.js
