+++ src/static/js/manage_mappings_utils.js
