+++ src/templates/manage_mappings.html
