+++ src/templates/supervisor_dashboard.html
