+++ tests/test_health.py
@@ -56,7 +56,7 @@ class TestHealthEndpoints:
         assert 'python_version' in data
         assert 'platform' in data
 
-    @patch('wkndPlanning.routes.health.current_app')
+    @patch('src.routes.health.current_app')
     def test_debug_endpoint_in_production(self, mock_app, client):
         """Test debug endpoint is blocked in production mode."""
         mock_app.config.get.return_value = False  # FLASK_DEBUG = False
@@ -68,7 +68,7 @@ class TestHealthEndpoints:
         assert 'error' in data
         assert 'not available in production' in data['error']
 
-    @patch('wkndPlanning.services.health_check.HealthChecker.check_database_health')
+    @patch('src.services.health_check.HealthChecker.check_database_health')
     def test_health_check_unhealthy_database(self, mock_db_check, client):
         """Test health check when database is unhealthy."""
         mock_db_check.return_value = (False, "Database connection failed")
@@ -85,7 +85,7 @@ class TestHealthChecker:
 
     def test_database_health_check_success(self, app):
         """Test successful database health check."""
-        from wkndPlanning.services.health_check import HealthChecker
+        from src.services.health_check import HealthChecker
 
         with app.app_context():
             checker = HealthChecker()
@@ -96,7 +96,7 @@ class TestHealthChecker:
 
     def test_filesystem_health_check_success(self, app):
         """Test successful filesystem health check."""
-        from wkndPlanning.services.health_check import HealthChecker
+        from src.services.health_check import HealthChecker
 
         with app.app_context():
             checker = HealthChecker()
@@ -107,7 +107,7 @@ class TestHealthChecker:
 
     def test_configuration_health_check_success(self, app):
         """Test successful configuration health check."""
-        from wkndPlanning.services.health_check import HealthChecker
+        from src.services.health_check import HealthChecker
 
         with app.app_context():
             checker = HealthChecker()
@@ -118,7 +118,7 @@ class TestHealthChecker:
 
     def test_get_application_metrics(self, app, test_db):
         """Test application metrics collection."""
-        from wkndPlanning.services.health_.health_check import HealthChecker
+        from src.services.health_.health_check import HealthChecker
 
         with app.app_context():
             checker = HealthChecker()
